+++ ndk-cache-nostr/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @nostr-dev-kit/ndk-cache-nostr
 
+## 0.1.30
+
+### Patch Changes
+
+- Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.18
+
+## 0.1.29
+
+### Patch Changes
+
+- Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.17
+
+## 0.1.28
+
+### Patch Changes
+
+- Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.16
+
 ## 0.1.27
 
 ### Patch Changes