+++ ndk-cache-redis/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @nostr-dev-kit/ndk-cache-redis
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.18
+
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.17
+
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.16
+
 ## 0.1.13
 
 ### Patch Changes