+++ ndk-core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @nostr-dev-kit/ndk
 
+## 2.14.18
+
+### Patch Changes
+
+- bump
+
+## 2.14.17
+
+### Patch Changes
+
+- bump
+
+## 2.14.16
+
+### Patch Changes
+
+- export NDKFollowPack
+
 ## 2.14.15
 
 ### Patch Changes